refactor(lamp): extract Hue URI authorization
